+++ packages/web/.storybook/main.mjs
@@ -7,7 +7,7 @@ function getAbsolutePath(packageName) {
 
 /** @type { import('@storybook/react-vite').StorybookConfig } */
 export default {
-  framework: '@storybook/react-vite',
+  framework: getAbsolutePath('@storybook/react-vite'),
   stories: ['../**/*.stories.@(js|jsx|mjs|ts|tsx)'],
   typescript: {
     reactDocgen: false,
@@ -17,7 +17,7 @@ export default {
   },
   addons: [
     {
-      name: '@storybook/addon-links',
+      name: getAbsolutePath('@storybook/addon-links'),
       options: { docs: false }, // no mdx
     },
     '@storybook/addon-links',
