+++ packages/web/app/api/mutations/index.js
@@ -6,3 +6,4 @@ export * from './usePatientMove';
 export * from './usePatientPaymentsMutation';
 export * from './useNotificationsMutation';
 export { useUserPreferencesMutation } from './useUserPreferencesMutation';
+export { useUpdateProgramRegistryMutation } from './useUpdateProgramRegistryMutation';
